Added Swarm

+++ src/main/java/ru/itis/swarm/FitnessFunction.java
@@ -1,5 +1,16 @@
 package ru.itis.swarm;
 
+/**
+ * Функциональный интерфейс вычисления уровня приспособленности
+ */
 public interface FitnessFunction {
-	double getFitness(long[] particlePosition);
+
+    /**
+     * Получить текущий уровень приспособленности частицы
+     *
+     * @param particlePosition Координаты частицы
+     * @return числовое значение уровня
+     */
+    Double getFitness(Double[] particlePosition);
+
 }
